requirements.txt
langchain==0.1.11
langchain-community==0.0.27
python-dotenv
flask
PyMuPDF